+++ FA2sp/Ext/CMapData/Hooks.cpp
@@ -4,6 +4,9 @@
 
 #include "../../FA2sp.h"
 #include "../../Helpers/TheaterHelpers.h"
+#include "../../Helpers/STDHelpers.h"
+
+#include <Miscs/Miscs.h>
 
 DEFINE_HOOK(424421, CFinalSunDlg_OnInitDialog_InitOreValue, 5)
 {
@@ -108,3 +111,67 @@ DEFINE_THEATER_NAME_FIX(49E863, LUNAR, EDI);
 DEFINE_THEATER_NAME_FIX(49EAA4, DESERT, ESI);
 
 #undef DEFINE_THEATER_NAME_FIX
+
+DEFINE_HOOK(4A4AA6, CMapData_UpdateMapFieldData_Structures_RenderData_Clear, 6)
+{
+    CMapDataExt::BuildingRenderDatasFix.clear();
+    return 0;
+}
+
+DEFINE_HOOK(4A4C8E, CMapData_UpdateMapFieldData_Structures_RenderData_Load, 6)
+{
+    GET(ppmfc::CString*, pString, EAX);
+
+    const auto& splits = STDHelpers::SplitString(*pString);
+    
+    BuildingRenderData data;
+    data.HouseColor = Miscs::GetColorRef(splits[0]);
+    data.ID = splits[1];
+    data.Strength = std::clamp(atoi(splits[2]), 0, 256);
+    data.Y = atoi(splits[3]);
+    data.X = atoi(splits[4]);
+    data.Facing = atoi(splits[5]);
+    data.PowerUpCount = atoi(splits[10]);
+    data.PowerUp1 = splits[12];
+    data.PowerUp2 = splits[13];
+    data.PowerUp3 = splits[14];
+    CMapDataExt::BuildingRenderDatasFix.push_back(data);
+
+    R->ESP(R->ESP() + 0xC);
+    return 0x4A4F6D;
+}
+
+DEFINE_HOOK(4A4F7A, CMapData_UpdateMapFieldData_Structures_SKIPCOPY, 9)
+{
+    return 0x4A4F83;
+}
+
+DEFINE_HOOK(4A57CD, CMapData_UpdateMapFieldData_Structures_SKIPDTOR, F)
+{
+    return 0x4A5817;
+}
+
+DEFINE_HOOK(4C3C20, CMapData_GetStructureRenderData, 6)
+{
+    GET_STACK(int, ID, 0x4);
+    GET_STACK(CBuildingRenderData*, pRet, 0x8);
+
+    if (ID >= 0 && static_cast<unsigned int>(ID) < CMapDataExt::BuildingRenderDatasFix.size())
+    {
+        const auto& data = CMapDataExt::BuildingRenderDatasFix[ID];
+        *reinterpret_cast<unsigned int*>(&pRet->HouseColor) = data.HouseColor;
+        pRet->ID = data.ID;
+        pRet->Strength = static_cast<unsigned char>(data.Strength);
+        pRet->X = data.X;
+        pRet->Y = data.Y;
+        pRet->Facing = data.Facing;
+        pRet->PowerUpCount = data.PowerUpCount;
+        pRet->PowerUp1 = data.PowerUp1;
+        pRet->PowerUp2 = data.PowerUp2;
+        pRet->PowerUp3 = data.PowerUp3;
+
+        CMapDataExt::CurrentRenderBuildingStrength = data.Strength;
+    }
+
+    return 0x4C3CCC;
+}